requirements.txt
# Role-Aware Financial Assistant - Dependencies
# Python 3.9+

# Core LangChain packages
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0

# Vector Store (FAISS - no Rust required)
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.0

# OpenAI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
numpy<2.0.0

# UI (using 3.x for Python 3.9 compatibility)
gradio>=3.50.0,<4.0.0

# Vision RAG (CLIP - lightweight, ~400MB)
matplotlib>=3.5.0
pillow>=9.0.0
# Note: CLIP comes with transformers (already installed via sentence-transformers)
